starter/requirements.txt
tqdm
click
PyYAML
opencv-python==4.6.0.66
albumentations==1.3.1
pytorch_toolbelt
segmentation_models_pytorch
torch==1.11.0
ray[tune]
aioredis==1.3.1
aiohttp==3.8.0
s3fs==2021.10.1
matplotlib
smart_open
setuptools==59.5.0
h5py
scipy
pyyaml
pytest
pytorch_lightning
notebook